requirements.txt
flask==3.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.31.0
flask-cors==4.0.0
python-dotenv==1.0.0
flask-socketio==5.3.6
eventlet==0.33.3
flask-talisman==1.1.0
